requirements.txt
python-telegram-bot
openai
python-dotenv
beautifulsoup4
numpy
unidecode
pyspellchecker
